Refactor controller for 'Model property' to be extensible for other ways to get the properties, but this is a partial solution as ModelPropertyResolver would need a similar change #535
